+++ CHANGELOG.md
@@ -127,6 +127,7 @@
 - Method `createAttribute` of Pim/Bundle/CatalogBundle/Manager/AttributeManager.php is now deprecated use `AttributeFactory::createAttribute` instead
 - Constructor of `Pim\Bundle\EnrichBundle\Controller\ChannelController` now takes a BulkSaver as last argument (to save locales)
 - Constructor of `Pim\Bundle\CatalogBundle\Manager\AttributeManager` has been changed
+- Constructor of `Pim\Bundle\CatalogBundle\Manager\AttributeGroupManager` has been changed
 
 # 1.3.x