+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# 1.3.x
+# 1.3.15 (2015-06-05)
 
 ## Bug fixes
 - PIM-4308: MongoDB indexes are removed on schema update